+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 #### Upcoming Changes
 
+* feat: implement `kzg` data availability hints [#1887](https://github.com/lambdaclass/cairo-vm/pull/1887)
+
 #### [2.0.0-rc2] - 2024-12-12
 
 * fix: Change wildcard getrandom dependency.